// src/generated/models/Dev_table_programsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dev_table_programsdev_level = {
  775730000: 'CarreraUniversitaria',
  775730001: 'Especializaci_n',
  775730002: 'Maestr_a',
  775730003: 'EntrenamientosMedicosAvanzados',
  775730004: 'Doctorados',
  775730005: 'CursosPreuniversitarios',
  775730006: 'Posdoctorados',
  775730007: 'Educaci_ncontinua'
} as const;
export type Dev_table_programsdev_level = keyof typeof Dev_table_programsdev_level;
export const Dev_table_programsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dev_table_programsstatecode = keyof typeof Dev_table_programsstatecode;
export const Dev_table_programsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dev_table_programsstatuscode = keyof typeof Dev_table_programsstatuscode;

export interface Dev_table_programsBase {
  dev_level?: Dev_table_programsdev_level;
  dev_nameprogram?: string;
  "dev_table_faculty@odata.bind": string;
  dev_table_programid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dev_table_programsstatecode;
  statuscode?: Dev_table_programsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dev_table_programs extends Dev_table_programsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dev_levelname?: string;
  dev_table_facultyname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dev_table_faculty?: object;
  _dev_table_faculty_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}